Refactor to avoid `rates_data.csv`
Instead of downloading and saving to rates_data.csv,
the script will read directly from url. This requires Perl,
but simplifies the program and means it is up-to-date each time
the app is run.

It is reasonable to rely upon Perl being installed.
Also cleaned up code a bit: spacing, comments, etc.